+++ source/ada/lsp-ada_driver.adb
@@ -25,6 +25,7 @@ with GNAT.Strings;
 
 pragma Warnings (Off, "is an internal GNAT unit");
 with Gnatformat.Configuration;
+with LSP.Ada_Handlers.Open_ALS_Log_File_Commands;
 with System.Soft_Links;
 with System.Secondary_Stack;
 
@@ -56,6 +57,7 @@ with LSP.Ada_Handlers.Mains_Commands;
 with LSP.Ada_Handlers.Named_Parameters_Commands;
 with LSP.Ada_Handlers.Object_Dir_Commands;
 with LSP.Ada_Handlers.Other_File_Commands;
+with LSP.Ada_Handlers.Open_Project_File_Commands;
 with LSP.Ada_Handlers.Project_File_Commands;
 with LSP.Ada_Handlers.Project_Reload_Commands;
 with LSP.Ada_Handlers.Refactor.Add_Parameter;
@@ -175,6 +177,10 @@ procedure LSP.Ada_Driver is
         (LSP.Ada_Handlers.Suspend_Executions.Suspend_Execution'Tag);
       LSP.Ada_Commands.Register
         (LSP.Ada_Handlers.Project_Reload_Commands.Command'Tag);
+      LSP.Ada_Commands.Register
+        (LSP.Ada_Handlers.Open_Project_File_Commands.Command'Tag);
+      LSP.Ada_Commands.Register
+        (LSP.Ada_Handlers.Open_ALS_Log_File_Commands.Command'Tag);
       LSP.Ada_Commands.Register
         (LSP.Ada_Handlers.Show_Dependencies_Commands.Command'Tag);
       LSP.Ada_Commands.Register